requirement.txt
# Core Web Framework
fastapi==0.117.1
uvicorn==0.37.0
starlette==0.48.0
python-multipart==0.0.20

# Data Processing
pandas==2.3.2
numpy==1.26.4
openpyxl==3.1.5
pyarrow==21.0.0

# Database
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
pymysql==1.1.2

# AI/ML Libraries
groq==0.37.1
distro==1.9.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langchain-core==1.1.0
langsmith==0.4.49
great-expectations==0.18.22
scikit-learn==1.7.2
scipy==1.15.3

# Streamlit UI
streamlit==1.50.0
altair==4.2.2
pydeck==0.9.1

# HTTP and Async
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
h11==0.16.0
httplib2==0.31.0
requests==2.32.5
requests-toolbelt==1.0.0

# Environment and Config
python-dotenv==1.2.1
pydantic==2.11.9
pydantic_core==2.33.2
annotated-types==0.7.0

# Jupyter/Notebook (optional - remove if not needed)
jupyterlab==4.4.9
notebook==7.4.7
ipython==8.12.0
ipykernel==6.30.1
ipywidgets==8.1.7

# Data Validation & Serialization
marshmallow==3.26.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.11.4
ormsgpack==1.12.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tzlocal==5.3.1
arrow==1.3.0
click==8.3.0
toml==0.10.2
PyYAML==6.0.3
packaging==25.0
setuptools==80.9.0
six==1.17.0
tenacity==9.1.2
tqdm==4.67.1

# gRPC
grpcio==1.76.0
grpcio-status==1.71.2
proto-plus==1.26.1
protobuf==5.29.5

# Cryptography & Security
cryptography==46.0.1
certifi==2025.8.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
cffi==2.0.0
pycparser==2.23

# Git Integration
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# Template & Markup
Jinja2==3.1.6
MarkupSafe==3.0.3
beautifulsoup4==4.14.0
soupsieve==2.8
bleach==6.2.0
tinycss2==1.4.0
webencodings==0.5.1

# Terminal & Process
psutil==7.1.0
pexpect==4.9.0
ptyprocess==0.7.0

# Misc
attrs==25.3.0
cachetools==6.2.0
charset-normalizer==3.4.3
colorama==0.4.6
idna==3.10
urllib3==2.5.0
sniffio==1.3.1
anyio==4.11.0
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
propcache==0.4.1
xxhash==3.6.0
zstandard==0.25.0
uritemplate==4.2.0
typing_extensions==4.15.0
typing-inspection==0.4.1

# Jupyter Dependencies
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.0
async-lru==2.0.5
babel==2.17.0
blinker==1.9.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
joblib==1.5.2
json5==0.12.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
lark==1.3.0
makefun==1.16.0
matplotlib-inline==0.1.7
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-json-logger==3.3.0
pyzmq==27.1.0
referencing==0.36.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
ruamel.yaml==0.17.40
Send2Trash==1.8.3
stack-data==0.6.3
terminado==0.18.1
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
uri-template==1.3.0
wcwidth==0.2.14
webcolors==24.11.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
appnope==0.1.4
greenlet==3.2.4
